Superior Weight Logic
Most weight-based approaches only assign a positive weight to a very small subset of overall training observations. For our computations only these observations with positive weight are of interest. By incorporating this insight, we can speed up the calculation of decisions by usually at least one magnitude.

On top of that, this code logic avoids running into RAM issues when working with datasets with >500k rows, because in that case the weight-array would become unnecessarily large.